Random Quote Generator
This is my attempt at the random quote generator project assigned to students as part of the curriculum of FreeCodeCamp. Currently it is limited as to how many quotes it can generate because I have not assigned an API to it. Be on the lookout for a newer version in the near future.